move tests to rdffile tests, re #3706
